Data_Engineering/requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
beautifulsoup4
boto3